[SDK] Fix dtype mismatch in `test_attention_no_opaque.py`
This PR casts `hidden_state` to `float32`. Without this fix, we get a
`ValueError: DType mismatch: expected f32 at position 0 but got f64
instead.`

MAX_INTEGRATION_TESTS_REV_ID: 01e48afd6ee809e5d853694a23e2f8ac67fb5a7e

+++ max/tests/integration/pipelines/python/nn/attention/test_attention_no_opaque.py
@@ -232,7 +232,9 @@ def test_compare_attention_with_rope_no_opaque() -> None:
             input_row_offsets.tensor,
         )
 
-    hidden_state = np.random.randn(max_seq_len * batch_size, dim)
+    hidden_state = np.random.randn(max_seq_len * batch_size, dim).astype(
+        np.float32
+    )
     input_row_offsets = np.array(
         [0, max_seq_len, max_seq_len * 2], dtype=np.uint32
     )
